# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VETRA API - Platform Kesehatan Hewan Peliharaan",
    "description": "Dokumentasi Lengkap API untuk VETRA - Platform Booking Klinik, Konsultasi Dokter, Chatting, dan AI Chatbot",
    "version": "1.0.0",
    "contact": {
      "name": "Vetra API Support",
      "email": "support@vetra.com"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000/api",
      "description": "Development Server"
    }
  ],
...
  "paths": {
...
    "/doctors": {
      "get": {
        "tags": ["Public Content"],
        "summary": "Mendapatkan daftar dokter",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List data dokter" }
        }
      }
    },
...
    "/clinics": {
      "get": {
        "tags": ["Public Content"],
        "summary": "Mendapatkan daftar klinik",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } }
        ],
        "responses": {
          "200": { "description": "List data klinik" }
        }
      }
    },
...
    "/articles": {
      "get": {
        "tags": ["Public Content"],
        "summary": "Mendapatkan daftar artikel",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Daftar artikel kesehatan hewan" }
        }
      }
    },
...
    "/doctor/dashboard": {
      "get": {
        "tags": ["Doctor Panel"],
        "summary": "Dashboard statistik dokter",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": { "description": "Statistik & recent bookings" }
        }
      }
    },
...
    "/clinic/dashboard": {
      "get": {
        "tags": ["Clinic Panel"],
        "summary": "Dashboard statistik klinik",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": { "description": "Statistik global klinik" }
        }
      }
    },
    "/clinic/bookings": {
      "get": {
        "tags": ["Clinic Panel"],
        "summary": "Mengambil data booking klinik",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": { "description": "List data booking klinik" }
        }
      }
    },
    "/clinic/doctors": {
      "get": {
        "tags": ["Clinic Panel"],
        "summary": "Mengambil dokter klinik tersebut",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": { "description": "List dokter klinik" }
        }
      },
      "post": {
        "tags": ["Clinic Panel"],
        "summary": "Menambahkan dokter ke klinik",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string", "example": "Dr. Sarah" },
                  "email": { "type": "string", "format": "email", "example": "sarah@vetra.id" },
                  "password": { "type": "string", "example": "password123" },
                  "phone": { "type": "string", "example": "08123445" },
                  "spesialis": { "type": "string", "example": "Dokter Umum" },
                  "experience_years": { "type": "integer", "example": 3 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Dokter berhasil ditambahkan" }
        }
      }
    },
...
    "/admin/users": {
      "get": {
        "tags": ["Admin Panel"],
        "summary": "List seluruh user terdaftar",
        "security": [{"bearerAuth": []}],
        "parameters": [
          { "name": "role", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List data users" }
        }
      },
      "post": {
        "tags": ["Admin Panel"],
        "summary": "Membuat user baru via Admin",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password", "role"],
                "properties": {
                  "name": { "type": "string", "example": "New User Admin" },
                  "email": { "type": "string", "format": "email", "example": "newadmin@vetra.id" },
                  "password": { "type": "string", "example": "password123" },
                  "role": { "type": "string", "enum": ["user", "doctor", "clinic", "admin"], "example": "user" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User berhasil dibuat" }
        }
      }
    },
...
    "/admin/messages": {
      "get": {
        "tags": ["Admin Panel"],
        "summary": "List pesan bantuan (inbox)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["unread", "read", "replied"] } }
        ],
        "responses": {
          "200": { "description": "List inbox" }
        }
      }
    },
...
    "/admin/articles": {
      "get": {
        "tags": ["Admin Panel"],
        "summary": "List artikel admin",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": { "description": "List artikel" }
        }
      },
      "post": {
        "tags": ["Admin Panel"],
        "summary": "Membuat artikel baru",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "slug", "content"],
                "properties": {
                  "title": { "type": "string", "example": "Cara Merawat Anak Kucing" },
                  "slug": { "type": "string", "example": "cara-merawat-anak-kucing" },
                  "excerpt": { "type": "string", "example": "Langkah awal merawat anak kucing" },
                  "content": { "type": "string", "example": "Lorem ipsum..." },
                  "status": { "type": "string", "enum": ["published", "draft"], "example": "published" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Artikel berhasil dibuat" }
        }
      }
    },
...
    "/v2/doctors": {
      "get": {
        "tags": ["API Key Integration (v2)"],
        "summary": "List dokter (API Key)",
        "security": [{"apiKeyAuth": []}],
        "responses": {
          "200": { "description": "List data dokter" }
        }
      }
    },
    "/v2/clinics": {
      "get": {
        "tags": ["API Key Integration (v2)"],
        "summary": "List klinik (API Key)",
        "security": [{"apiKeyAuth": []}],
        "responses": {
          "200": { "description": "List data klinik" }
        }
      }
    },
    "/v2/articles": {
      "get": {
        "tags": ["API Key Integration (v2)"],
        "summary": "List artikel (API Key)",
        "security": [{"apiKeyAuth": []}],
        "responses": {
          "200": { "description": "List data artikel" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Autentikasi standar JWT. Masukkan token Anda (Tanpa menulis 'Bearer ')"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Autentikasi integrasi API Key statis. Kirimkan token API Key Anda pada header X-API-Key"
      }
    },
...
  }
}